fix: rename more files, remove tests
