--- android/app/src/main/kotlin/com/example/notice_board/MainActivity.kt
@@ -1,6 +1,0 @@
-package com.example.notice_board
-
-import io.flutter.embedding.android.FlutterActivity
-
-class MainActivity: FlutterActivity() {
-}
