+++ android/app/src/main/kotlin/com/example/noticeboard/MainActivity.kt
@@ -1,0 +1,5 @@
+package com.example.noticeboard
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity()
