+++ lib/views/mainPages/settings_screen.dart
@@ -1,5 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:flutter/src/widgets/framework.dart';
 
 class SettingsScreen extends StatefulWidget {
   const SettingsScreen({super.key});
