+++ pubspec.yaml
@@ -32,7 +32,7 @@ dependencies:
   flutter:
     sdk: flutter
   path: any
-  intl: ^0.18.0
+  intl: ^0.19.0
   url_launcher: ^6.1.9
   google_nav_bar: ^5.0.6
 
@@ -42,11 +42,11 @@ dependencies:
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
   get: ^4.6.5
-  http: ^0.13.5
+  http: ^1.2.1
   json_serializable: ^6.6.1
   flutter_dotenv: ^5.0.2
   webview_flutter: ^4.0.5
-  flutter_launcher_icons: ^0.12.0
+  flutter_launcher_icons: ^0.13.1
   cached_network_image: ^3.2.3
   shared_preferences: ^2.1.0
   provider: ^6.0.5
@@ -62,7 +62,7 @@ dev_dependencies:
   # activated in the `analysis_options.yaml` file located at the root of your
   # package. See that file for information about deactivating specific lint
   # rules and activating additional ones.
-  flutter_lints: ^2.0.0
+  flutter_lints: ^3.0.1
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
@@ -71,7 +71,6 @@ dev_dependencies:
 flutter:
   assets:
     - images/mitBuilding.jpeg
-    - .env
     - images/placeholder.jpg
     - images/loading.gif
     - images/cxlogo.png
